Move deviceListItem and recoverMethods to a component (#1274)
* move deviceListItem and recoverMethods to a component

* remove unused imports

* moves devices Section to its own component

* moves devices Section to its own component

* make naming of components more consistant

* add changes from merge to specific components
